move operators to where they are used
