+++ utils/__init__.py
@@ -138,7 +138,7 @@ utils_modules = [
     # UI operators and tools
     "sv_IO_pointer_helpers",
     "sv_IO_monad_helpers", "sv_operator_utils", "sv_IO_panel_operators",
-    "sv_panels_tools", "sv_gist_tools", "sv_IO_panel_tools", "sv_load_archived_blend",
+    "sv_gist_tools", "sv_IO_panel_tools", "sv_load_archived_blend",
     "monad", "sv_help", "sv_default_macros", "sv_macro_utils", "sv_extra_search", "sv_3dview_tools",
     "sv_update_utils", "sv_obj_helper", "sv_batch_primitives", "sv_idx_viewer28_draw",
     "sv_texture_utils",
